refactor: move BaseButton to internal
