+++ .nx/version-plans/version-plan-1786644540000-ui-rnative.md
@@ -1,0 +1,5 @@
+---
+'@ledgerhq/lumen-ui-rnative': patch
+---
+
+refactor(Button): move BaseButton to internal components and derive Button/IconButton/MediaButton types
