+++ libs/ui-rnative/src/lib/Components/core/IconButton/IconButton.tsx
@@ -1,4 +1,4 @@
-import { BaseButton } from '../Button/BaseButton';
+import { BaseButton } from '../../internal/BaseButton';
 import type { IconButtonProps } from './types';
 
 export const IconButton = (props: IconButtonProps) => {
