+++ libs/ui-rnative/src/lib/Components/internal/BaseButton/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseButton';
+export * from './types';
